Make the RemoteConnection private under RemoteConnections and document the setup
You’d never instantiate it on its own.
